change ShowMessageCount instance

+++ Growthbeat/GrowthPush/GrowthPush.m
@@ -430,7 +430,7 @@ const CGFloat kDefaultMessageInterval = 1.0f;
         }
         
         GPShowMessageCount *showMessageCount = [GPShowMessageCount receiveCount:[[self waitClient] id] applicationId:self.applicationId credentialId:self.credentialId taskId:message.task.id messageId:message.id];
-        [[self logger] info:[NSString stringWithFormat:@"Success show message (count: %ld, messageId: %@, taskId: %@)", showMessageCount.count, showMessageCount.messageId, showMessageCount.taskId]];
+        [[self logger] info:[NSString stringWithFormat:@"Success show message (count: %ld, messageId: %@, taskId: %@)", (long) showMessageCount.count, showMessageCount.messageId, showMessageCount.taskId]];
         
         break;
